+++ src/basic/decks/euchre24.rs
@@ -7,11 +7,13 @@ use crate::basic::types::basic_card::BasicCard;
 use crate::basic::types::pips::Pip;
 use crate::basic::types::traits::{Decked, DeckedBase};
 use crate::prelude::{Card, Pile};
+use alloc::string::{String, ToString};
+use alloc::vec::Vec;
 #[cfg(feature = "colored-display")]
 use colored::Color;
+use core::hash::Hash;
 #[cfg(feature = "colored-display")]
 use std::collections::HashMap;
-use std::hash::Hash;
 
 /// This deck represents the most common 24 card form of
 /// [Euchre](https://en.wikipedia.org/wiki/Euchre) with