+++ src/basic/decks/pinochle.rs
@@ -8,11 +8,13 @@ use crate::basic::types::traits::{Decked, DeckedBase};
 #[cfg(feature = "colored-display")]
 use crate::prelude::Standard52;
 use crate::prelude::{Card, Pile};
+use alloc::string::{String, ToString};
+use alloc::vec::Vec;
 #[cfg(feature = "colored-display")]
 use colored::Color;
+use core::hash::Hash;
 #[cfg(feature = "colored-display")]
 use std::collections::HashMap;
-use std::hash::Hash;
 
 #[derive(Clone, Copy, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub struct Pinochle {}